Clean up message handler
